app/requirements.txt
altgraph==0.17.4
importlib-metadata==7.0.1
joblib==1.3.2
macholib==1.16.3
mordred==1.2.0
networkx==2.8.8
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
python-dateutil==2.8.2
pytz==2024.1
pywebio==1.8.3
rdkit-pypi==2022.9.5
scipy==1.12.0
six==1.16.0
tornado==6.4
tqdm==4.66.1
tzdata==2023.4
ua-parser==0.18.0
user-agents==2.2.0
xgboost==2.0.3
zipp==3.17.0
